+++ site/lib/component-slugs.ts
@@ -7,7 +7,7 @@
  *  in DocsLayout.astro when adding a component. */
 export const COMPONENT_SLUGS = [
   'banner', 'button', 'card', 'checkbox', 'dialog', 'expander', 'icon', 'input',
-  'input-autocomplete', 'input-date', 'input-time',
+  'input-autocomplete', 'input-date', 'input-time', 'slider',
   'menu', 'progress', 'radio', 'select', 'select-faceted', 'stickers', 'table',
   'switch', 'tabs', 'tag', 'text', 'title', 'toaster', 'tooltip',
 ]